exclude embedding from get docs

+++ src/main/java/com/example/application/dtos/DocumentDTO.java
@@ -1,4 +1,4 @@
 package com.example.application.dtos;
 
-public record DocumentDTO(int id, String name, String content) {
+public record DocumentDTO(Long id, String name, String content) {
 }
